+++ projects/start-os/web/ui/src/app/routes/portal/routes/updates/item.component.ts
@@ -33,6 +33,7 @@ import {
   PackageDataEntry,
   UpdatingState,
 } from 'src/app/services/patch-db/data-model'
+import { MarketplaceAlertsService } from '../marketplace/services/alerts.service'
 import UpdatesComponent from './updates.component'
 
 @Component({
@@ -278,6 +279,7 @@ export class UpdatesItemComponent {
   private readonly service = inject(MarketplaceService)
   private readonly dialog = inject(DialogService)
   private readonly hiddenUpdates = inject(HiddenUpdatesService)
+  private readonly alerts = inject(MarketplaceAlertsService)
 
   readonly parent = inject(UpdatesComponent)
   readonly expanded = signal(false)
@@ -290,8 +292,17 @@ export class UpdatesItemComponent {
   readonly pending = input.required<boolean>()
 
   async update() {
-    const { id, version } = this.item()
+    const { id, version, preDownloadAlert } = this.item()
     const url = this.parent.current()?.url || ''
+
+    if (
+      !(await this.alerts.alertPreDownload(
+        preDownloadAlert,
+        this.local().stateInfo.manifest.version,
+      ))
+    ) {
+      return
+    }
 
     this.ready.set(false)