+++ shared-libs/ts-modules/shared/src/i18n/dictionaries/pl.ts
@@ -821,4 +821,5 @@ export default {
   929: 'Malejąco',
   930: 'Serwisy z tego katalogu są utrzymywane przez społeczność Start9 i są w fazie testów beta. Należy spodziewać się błędów. Instalujesz je na własne ryzyko.',
   931: 'Wszystkie adresy IP, na które ta nazwa jest rozwiązywana na tej bramie, są wyłączone',
+  932: 'Zaczekaj!',
 } satisfies i18n